Updated PersistanceContext

+++ src/database/Persistent.java
@@ -1,5 +1,7 @@
 package database;
 
+import java.util.Objects;
+
 public interface Persistent {
-    public long getPersistanceId();
+    public int getPersistanceId();
 }
